requirements/pypi.txt
# Django
django

# Social Login
python-social-auth
social-auth-app-django

# Rest
djangorestframework
